More work on updated persistence implementation. No more build errors or warnings related to loadMeals() and saveMeals() methods, but persistence is not working. Suggested avenue of inquiry is in details.
DETAILS:
More work on loadMeals() and saveMeals(). Implemented new load and save methods with (shitty) error handling. No build issues, but persistence doesn't work still. Save gets called, confirmed. Working on getting debug error stream into Xcode when application is quit and relaunched. Old methods are still included in comments if turning this into a shippable build is necessary.
Suspect problem is with load method, and archive url only gets us to the folder where the data is stored, not the data itself

ISSUES:
IN#3, IN#4: in progress