Add limited adaptive behaviour to table view cells
DETAILS:
Add constraints to storyboard to clear ambiguity and adaptively shrink meal name label in Your Meals view
Test behaviour when cell shrinks and conclude that, in lieu of a truly adaptively-sizeable rating control, having it slide off the end of the cell is way less ugly than just scaling the width of the last star. Conclude that adapting this to smaller screens is probably best done with size classes. IN#2 resolved, IN#8 introduced.

ISSUES:
IN#2 resolved - table view cell adaptive layout
IN#8 introduced - Rating control falls off the end of the cell in small-width (320 pt wide) devices