Prototype a "not applicable" rating to demonstrate interface behaviour for client
DETAILS:
Add "not applicable" button to the rating control array
Set behaviour, clearing all stars when "not applicable" is selected.
Display this control to the user, highlighting the "N/A" title text when "not applicable" is selected
Note that saving will not work in this build when N/A is selected, as the save method checks for a rating <= 5 and N/A is implemented as a six
Better implementation would have notApplicableIsSelected as a flag and set the rating to zero when user selects N/A
Check that client was happy with the behaviour of the rating control interface

Note/story/reason: Mum, who works in education, was explaining how the highly paid govt. contractors who write their reporting software took an unreasonable amount of time and hand-wringing to implement a rating control that cleared the star rating when N/A is selected. I explained that the feature was so simple that I could implement it this afternoon to her satisfaction, despite less than three weeks of iOS development experience. I was correct. This took less than an hour to do, and mum confirmed that the control does exactly what she expects.

ISSUES:
nil