Deal with table view layout-while-not-visible warning, despite concluding that the warning was probably harmless :)
DETAILS:
Research warning, evidence points to warning being an overzealous suggestion. Poor performance not likely, and the warning is apparently repeated in Apple's own templates and tutorials.
Found a neat solution in moving the update to the viewDidAppear(_:) method of the table view controller. However, this calls a reload every time we return from a loaded meal. I can't imagine there's a huge performance difference, but hey, the warning's gone.

ISSUES:
IN#5: Table view bad practice: fixed! Closed and resolved.